freed split:

+++ srcs/stack_init.c
@@ -6,7 +6,7 @@
 /*   By: rafaelfe <rafaelfe@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/04 20:13:11 by rafaelfe          #+#    #+#             */
-/*   Updated: 2025/01/06 19:56:05 by rafaelfe         ###   ########.fr       */
+/*   Updated: 2025/01/07 13:04:53 by rafaelfe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ t_stack *new_node(int data);
 void	ft_stack_addback(t_stack **lst, t_stack *new);
 t_stack	*ft_stacklast(t_stack *lst);
 
-void	stack_init(t_stack **stack_a, int *num_array, int stack_size)
+void	stack_init(t_stack **stack_a, long *num_array, int stack_size)
 {
 	t_stack *node;
 	int		i;
